Create Get-ExchangeAdminAuditLogConfig.ps1 (#404)
* Create Get-ExchangeAdminAuditLogConfig.ps1

Discover Exchange Admin Audit log is enabled and admin and user log activities are logged. Output Status if configuration is enabled.

close #382

* Update Get-ExchangeAdminAuditLogConfig.ps1

* Update ExchangeDiscovery.psm1

updated script to call new function Get-ExchangeAdminAuditLogConfig

* Update ExchangeDiscovery.psd1

added Get-ExchangeAdminAuditLogConfig.ps1 to the Nested Modules

* Update Get-ExchangeAdminAuditLogConfig.ps1

corrected issue that prevented the output from being displayed. The $discoveredAdminAuditLogConfig did not need to be updated by the $exchangeAdminAuditLogConfig but simply have the object itself in the output.
Replaced $exchangeAdminAuditLogConfig with $discoveredAdminAuditLogConfig. to resolve

* Update Get-ExchangeAdminAuditLogConfig.ps1

didn't need array since there is only one output of AdminAuditLogEnabled. I kept it as the $discoveredAdminAuditLogConfig to match formatting of other scripts in EDU